feat: implement the services implementations and repositories

+++ src/main/java/com/ferrinsa/fairpartner/expense/repository/ExpenseRepository.java
@@ -3,6 +3,7 @@ package com.ferrinsa.fairpartner.expense.repository;
 import com.ferrinsa.fairpartner.expense.model.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -16,6 +17,6 @@ public interface ExpenseRepository extends JpaRepository<Expense,Long> {
 
     List<Expense> findByDateBetween(LocalDate startDate, LocalDate endDate);
 
-    List<Expense> findByAmountBetween(Double min,Double max);
+    List<Expense> findByAmountBetween(BigDecimal min, BigDecimal max);
 
 }
